Separate out octobot stuff from http stuff from `main` stuff
